byte -> short

+++ src/main/java/network/IMessageConsumer.java
@@ -2,5 +2,5 @@ package network;
 
 public interface IMessageConsumer<T> {
 
-    void deliverMessage(byte msgCode, T msg, Host from);
+    void deliverMessage(short msgCode, T msg, Host from);
 }
